+++ src/Application.php
@@ -38,8 +38,8 @@ class Application extends BaseApplication
         $commands[] = new Command\CleanTorrentsCommand();
         $commands[] = new Command\ListTorrentsCommand();
         $commands[] = new Command\RemoveDuplicatesCommand();
-        $commands[] = new Command\SendMetricsCommand();
 
+        $commands[] = new Command\StatsSend();
         $commands[] = new Command\WeburgDownload();
 
         $commands[] = new BashCompletion\CompletionCommand();